Moved the table header component to a table directory
